Automated the generation of many-to-many table names, uni and bi directional.
Bi directional many-to-many's now attempt to find the other side of the relationship and set both sides to be the same thing; if one side is explicitly set in the mapping then the other side will inherit it, if none are set the one will be generated. It will locate the other side by finding any many-to-many's on the child entity that are a collection of the parent, and optionally look at the property names if there is more than one many-to-many of the same type.

The only situation it can't cope with is if you have more than one many-to-many of the same type on the same two entities, with one of them not having a logically discernable.

Added a ManyToManyTableNameConvention base class that you can inherit from to make supplying your own formatting easier; all the pairing is already done for you if you use this base class, all you need to do is implement the abstract naming methods.